yooop

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: aragragu <aragragu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 17:29:04 by aragragu          #+#    #+#             */
-/*   Updated: 2023/11/09 19:32:58 by aragragu         ###   ########.fr       */
+/*   Updated: 2023/12/01 18:29:50 by aragragu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
